+++ src/theme.js
@@ -8,18 +8,18 @@ function $(selector, element = document) {
 
 function bindThemeSwitchers() {
   for (const element of $$("[data-theme-switcher]")) {
-    const c0 = element.dataset["bitTheme-0"];
-    const c1 = element.dataset["bitTheme-1"];
-    element.style.setProperty("--bit-color-0", c0);
-    element.style.setProperty("--bit-color-1", c1);
+    const c0 = element.dataset.bitTheme0;
+    const c1 = element.dataset.bitTheme1;
+    element.style.setProperty("--bit-color0", c0);
+    element.style.setProperty("--bit-color1", c1);
     element.addEventListener(
       "click",
       () => {
         for (const button of $$("[data-theme-switcher]")) {
           button.disabled = button === element;
         }
-        const c0 = element.dataset["bitTheme-0"];
-        const c1 = element.dataset["bitTheme-1"];
+        const c0 = element.dataset.bitTheme0;
+        const c1 = element.dataset.bitTheme1;
         setThemeColors(c0, c1);
         updateThemeExample(c0, c1);
       },
@@ -31,8 +31,8 @@ function bindThemeSwitchers() {
 function updateThemeExample(color0, color1) {
   $("#theme-example").textContent = `\
 .bit-root {
---bit-color-0: ${color0};
---bit-color-1: ${color1};
+--bit-color0: ${color0};
+--bit-color1: ${color1};
 }`;
 }
 
@@ -49,8 +49,8 @@ function restoreUserTheme() {
 
 function setThemeColors(color0, color1) {
   const root = $(":root");
-  root.style.setProperty("--bit-color-0", color0);
-  root.style.setProperty("--bit-color-1", color1);
+  root.style.setProperty("--bit-color0", color0);
+  root.style.setProperty("--bit-color1", color1);
   jsonStorageSet("user-theme", { color0, color1 });
 }