+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.0.4
+
+- Changes `--bit-color-0` to `--bit-color0`
+- Changes `--bit-color-1` to `--bit-color1`
+
 # v0.0.3
 
 - Fixes state for `:disabled:focus`